+++ assignments/03-STMS/wcs/src/events/PotEvent.cpp
@@ -1,4 +1,4 @@
 #include "events/PotEvent.hpp"
 
-PotEvent::PotEvent(float value) : Event(value) { }
+PotEvent::PotEvent(int value) : Event(value) { }
 EventType PotEvent::getType() const { return EventType::POT_EVT; }
